Corregir migraciones

+++ database/migrations/2025_11_17_170948_remove_technical_sheet_from_products_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('technical_sheet');
-        });
+        if (Schema::hasColumn('products', 'technical_sheet')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('technical_sheet');
+            });
+        }
     }
 
     /**
